Split services validate.js
